catfile: Fix ObjectReader not being dirty with in-use queue
In 21579e114 (catfile: Don't return readers to cache when their queues
are in use, 2023-03-02), we have fixed a bug where both the object info
and content reader were not considered dirty when their object queue was
still in use. This caused us to return the readers back to the cache
even though they might have still been in use.

Since then, we have introduced a new ObjectReader that combines both of
these other types into one. Unfortunately though, it seems like the
implementation was taken as an inspiration back when that bug still
existed, which has the consequence that the same bug exists in this new
type now.

Fix the bug by considering a reader to be dirty when its queue is in
use.

Changelog: fixed