catfile: Add mailmap support to new ObjectReader
In 0b032626, we have introduced mailmap support in our catfile package.
This change seems to have crossed with 950ca530, where we have
introduced a new catfile process type that uses `--batch-command` in
order to combine the `ObjectContentReader` and the `ObjectInfoReaders`
into a single process via `ObjectReader`. The new object reader does not
support mailmaps.

Add mailmap support for the new object reader, similar to the others.