localrepo: Tighten parsing of the HEAD reference
The `localrepo.HeadReference()` funciton is responsible for reading the
symbolic reference `HEAD`, returning whatever it contains. The function
is quite naive though and doesn't perform a lot of validation of the
output of git-symbolic-ref(1).

Refactor the code to read into a `strings.Builder` directly and perform
validation of that output.