git/log: Read last commit ID into buffer
The `log.LastCommitForPath()` function determines the last commit for a
specific path. This is done by first spawning git-log(1) with the path
as revision, where the returned commit ID is then passed used to read
the actual commit object.

The way we read the last commit ID is not exactly great though. We
manually read into a bytes buffer and don't perform any error checking
whatsoever -- we don't even call `Wait()` on the git-log(1) command, and
silently go ahead with the potentially-empty bytes buffer. Furthermore,
we don't verify whether the returned object ID matches our expectations
but just blindly pass it to git-cat-file(1).

Refactor the code to spawn git-log(1) with its stdout connected to a
`strings.Builder` and implementing both proper error checking and output
validation.