catfile: Refactor NotFoundError to match current best practices
The `catfile.NotFoundError` is not really following our current best
practices around error handling. It embeds a wrapped error that is the
same in all except one case, doesn't provide information about which
revision wasn't found and uses a `IsNotFound()` function instead of the
modern `errors.As()`. Last but not least, it's hard to construct this
error in dependents of this package, which is something we'll need in a
follow-up commit.

Refactor the error type to embed the revision that wasn't found instead
of an error such that we can attach structured error metadata to the
type. Furthermore, we convert all callsites to use `errors.As()` in
order to get rid of `IsNotFound()`.

Note that `catfile.GetTag()` abused this type by returning the error
with an embedded error effectively saying "object is not a tag". This
doesn't mean that the object wasn't found though, but instead indicates
that its type doesn't match our expectations, and thus using this error
is not a good fit. The function is converted to now use a plain error,
which is fine given than none of its callers actually checked whether
the returned error is of type `catfile.NotFoundError`.