smarthttp: Simplify ref advertisement
When handling the reference advertisement in our smarthttp service, we
manually copy output from Git into the writer. This is likely done this
way because we also write some preceding lines to that writer ourselves,
but is ultimately unnecessary.

Refactor the code to instead have the command write directly into the
writer to simplify the code and required error handling.