Suppress Run lint on packages testhelper imports
We cannot use `testhelper.Run` in packages that `testhelper` imports as
this would create a cyclic dependency. The tests in these packages also
rely on internal details and cannot be readily separated into `_test`
packages.

Add a `nolint:gitaly-linters` directive to the first file of the package
in anticipation of a new linter that catches packages missing execution
of `testhelper.Run`.