proxy: Create testhelper in proxy package
Currently the `testhelper_test.go` file is in the `proxy_test` package.
We need to add `TestMain`, but there are tests present in both `proxy`
and `proxy_test`. Adding it into `proxy_test` would lead to our
linter detecting that the `proxy` package is missing `TestMain`.

To resolve this, rename the existing `testhelper_test.go` to
`proxy_test_testhelper_test.go` and create a new testhelper in the
`proxy` package containing `TestMain`.