log: Remove global logging instance
We have refactored all uses of our global logging instance in favor of
using either an injected logger or the logger provided via the context.
Let's drop it and instead only configure the standard logger without
exposing it in any way. This ensures that dependencies which use it will
continue to function as expected.