git: Unify logic to pass dates to Git
There are multiple locations in our codebase where we need to pass dates
to Git, with the most common scenario being the `GIT_AUTHOR_DATE` and
`GIT_COMMITTER_DATE` environment variables. The way we do this is not
consistent though: in some locations we pass an RFC2822 date, whereas in
other locations we pass the Unix timestamp together with a timezone.

In general though, it is very much preferable to always pass RFC2822
dates. Git explicitly documents these as documented in git-commit(1) in
the "DATE FORMATS" section. And second, using Unix timestamps has the
disadvantage that Git only accepts Unix timestamps that have 8 or more
digits. It is thus impossible to pass any dates before 1970-04-26
17:46:40 +0000 UTC as Unix timestamp. And while this only rules out a
small window in time, this will certainly happen e.g. when the system
clock is not set up at all.

Introduce a new function `git.FormatTime()` that can be used to format
times using RFC2822. Convert callsites that used either RFC2822 or the
discouraged Unix time format to use that new function.