git/localrepo: Fix interpretation of monotonic clock as timezone
We have recently received a report where commit signatures written by
the `UserCherryPick()` RPC sometimes contain bogus timezones that look
like "Tue Nov 10 23:00:00 2009 +1234". This bug is caused by us passing
`time.Time.String()` as date to Git directly, which renders as a string
similar to "2009-11-10 23:00:00 +0200 UTC m=+1234.120000000". This is
not any format that is officially recognized by Git and noticably does
not match RFC2822.

What happens here is that Git interprets the "m=+1234.120000000" trailer
as timezone, which is obviously wrong. So far though, this bug only
surfaced in `UserCherryPick()`, even though the functionality is reused
by a lot of our RPCs. This is because `UserCherryPick()` (and also
`UserRevert()`) did not set up the client-side timestamp in the request.
Consequentially, Gitaly falls back to `time.Now()`, which will include
monotonic clock readings which are then rendered as that trailer.

While this bug in Rails has been fixed already, it's still wrong to pass
this date format to Git. So let's also fix Gitaly by using RFC2822
formatted dates instead.

Note that we unfortunately cannot add a test for this. This is because
Go does not expose a way to modify the monotonic clock readings, and
thus it's impossible to influence the "m=" part directly.

Changelog: fixed