git: Unify logic to format signature times
While the preceding commit has unified most callsites that pass
timestamps to Git, one notable ommission is the timestamps used when
writing raw author and committer signatures in Git, e.g. by calling
git-hash-objects(1). And while we should indeed be using RFC2822 as much
as possible to fix limitations when we would be passing Unix timestamps
instead, generating signature times is the one exception where we must
be using the Unix time.

Introduce a new function `git.FormatSignatureTime()` that does so.
Convert callsites that manually did this to use this function.