lint: Match bare function calls
Currently `Matcher` will only match against selector expressions, e.g.
`fmt.Println`. A bare function call, e.g. `Println`, will not be
matched.

Update `Matcher` to handle both package qualified and bare function
calls.

Also, fix the test comment for `NewUnavailable` to include `want`. Prior
to updating `Matcher` this condition was not being tested and was
silently failing. We leave off the URL at the end of the report for ease
of matching, this was being interpreted as a regexp by the `analysis`
package and is a pain to escape properly.