storagemgr: Fix flaky test when committing with cancelled context
We previously observe the following flaky test in our CI pipelines:

```
=== FAIL: internal/gitaly/storage/storagemgr TestPartitionManager/committing_transaction_failed_due_to_cancelled_context (0.42s)
    partition_manager_test.go:684:
        	Error Trace:	/builds/gitlab-org/gitaly/internal/gitaly/storage/storagemgr/partition_manager_test.go:684
        	Error:      	Target error should be in err chain:
        	            	expected: "context canceled"
        	            	in chain:
        	Test:       	TestPartitionManager/committing_transaction_failed_due_to_cancelled_context
```

What we're trying to do here is to check that context cancellation
propagates correctly and thus causes a transactional commit to fail.

The problem here though is that context cancellation and admission to
the transaction queue are part of the same select statement. Thus, if
both branches wouldn't block, then we have a 50% chance of picking
either of the branches and thus the test is indeterministic.

While we could fix this issue in production code by having an up front
check for a cancelled context, this wouldn't quite fix the underlying
race as the context can be cancelled at any point in time. Instead, we
thus only adapt the test to unset the admission queue. This has the
effect that we would now block indefinitely when trying to submit the
transaction, and thus we can test that the context cancellation would
cause us to bail out of that condition.