storagemgr: Introduce transaction manager factory into partition manager
In order to make one of our tests deterministic we will need to unset
the admission queue of the transaction manager such that submitting a
transaction would block indefinitely. Doing this after the transaction
manager has started already is producing races though, so we have to do
that at the time where the transaction manager is being created. This is
not really easy to do right now though because the test has no way to
hook into the creation of it.

Introduce a new transaction manager factory that allows us to hook into
the creation. While this factory shouldn't ever change for production
code, we will use it in the next commit to fix a test flake.