gitaly/config: Stop validating existence of GitLab secret file
In 5d7ecbde6 (gitaly/config: Allow configuration without `gitaly-hooks`
directory, 2023-08-24), we have introduced a that allows us to get rid
of the `[gitlab-shell]` section in favor of explicitly configured
options. As part of this change we also introduced validation of whether
the now explicitly configured GitLab secret file actually exists on
disk. As any setup that doesn't have the file is essentially broken we
thought this to be a good idea.

There was quite some fallout caused by this change though, mostly in
test setups where we don't care about whether Gitaly is fully functional
or not. So let's revert this change and stop verifying the existence of
the GitLab secret file.