git: Only set up receive-pack action for receive-pack hooks
We currently hardcode the hooks payload to always use the "receive-pack"
action. Both the reference-transaction and the packed-objects hooks
don't really have an associated action with them though. Let's set their
action to "unknown" such that we would notice eventually in case this
value ever bubbles up to the GitLab API.