log: Move testhelper_test.go to log_test package
Our linter will soon be able to detect if a blackbox test
project has tests but no `TestMain` in it or its primary package. In the
`internal/log` project we had to opt the primary package out of the
linter as it is imported by `internal/testhelper`, but this is not
necessary for `internal/log_test`.

Rename the existing `testhelper_test.go` file to
`log_testhelper_test.go` and create a new file in its place in the
`log_test` package that contains `TestMain`.