operations: Respect timezone info in signatures
There is a function `dateFromProto` in operations that assists in
generating signatures for new commits if nedded. However, this function
is not being applied in all relevant RPCs, leading to inconsistencies
with user's time zone.

This commit unifies the logic of generating date information for commit
signatures, thereby resolving this inconsistency.

Changelog: fixed
Signed-off-by: Xing Xin <xingxin.xx@bytedance.com>