repocleaner: Fix test race caused by too low liveness interval
One of our tests for the repocleaner checks that we skip over any
storages that are currently being processed by a concurrently running
cleaner already. This test is racy though because we frequently observe
that the second cleaner runs even though it should not.

This issue is caused by a liveness interval that's set quite low. This
interval dictates how often we update the liveness timestamp in the
acquired database lock -- and if a concurrent repocleaner sees that the
lock hasn't been updated since the last time period it'll evict the lock
and assume that the other repocleaner has died.

Of course, due to the busy nature of our CI runners, we frequently
exceed the liveness update period. So let's fix this by bumping it to
one hour instead of one second.