cli/gitaly: Make files implementing subcommands more discoverable
In our Praefect CLI, we implement each subcommand in a file that has a
"subcmd_" prefix. This helps us to make the respective implementations
easy to discover and clearly distinguishes "plumbing" functionality and
the respective subcommands.

In the Gitaly CLI we don't use the same pattern yet. Rename the files
such that we do.