Praefect: Log ignored sidechannel error during proxy
Since a prior commit, Praefect proxy always closes the sidechannel
connection. There is a possibility that the sidechannel's waiter returns
error when the connection is closed. When the upstream returns an error,
the error is propagated back to downstream. In such cases, the waiter
error is ignored.

This commit logs this type of error when it's swallowed. Most of the
changes are about passing a logger to the proxy intepretors.