gittest: Introduce repository executor that handles absolute paths
Many of our `gittest` helper functions work with absolute paths to the
repository instead of accepting a `gitalypb.Repository`. We're about to
introduce a new set of functions that reuse common functionality inside
of the `internal/git` package, which accepts a `git.RepositoryExecutor`.
This would create a mismatch between APIs provided by the `gittest`
package though.

Introduce a new `RepositoryPathExecutor` that implements the interface,
but can be constructed from an absolute path. That allows us to neatly
combine both worlds.