golangci-lint: Don't exclude error checks for `pool.Close()`
With our migration to use the internal pool functionality we have also
converted all callsites to use `testhelper.MustClose()` to close the
pool with proper error checking in place. Because of that, we only have
a single callsite left where we don't check the error code when closing
the pools, which thus makes our linter rule superfluous where we don't
require error checks for that function.

Remove the rule and add error checking to the last remaining callsite
that doesn't.