tests: Use `gittest.WriteBlob()` more extensively
Refactor tests to use `gittest.WriteBlob()` instead of the equivalent
function of the `localrepo` package. While this decreases the test
coverage of production-level code for now, this will be addressed in a
later commit where we reimplement `gittest` to use the same code as the
`localrepo` package. Meanwhile, the `gittest` function is easier to use
and less verbose.