gittest: Reimplement writing blobs to use production-level code
Reimplement the `WriteBlob()` and `WriteBlobs()` test helpers to reuse
our production-level code in `git.WriteBlob()`. This improves test
coverage of our production code and reduces code duplication.

Note that one of our tests needs some small fixup to create its
repository in a different location. This is because the usual safety
mechanisms now kick in with the new code, namely that spawned Git
commands actually run in a repository that is contained in the storage
root.