command: Reliably propagate read errors on context cancellation
On context cancellation, we will send a signal to the command's process
and then wait for it to exit. This means that the standard streams are
only getting closed _after_ we have sent a signal as part of `Wait()`.
Consequentially, we have a race between the process exiting due to the
signal and thus gracefully closing its standard streams, and ourselves
to close the streams and can thus lead to downstream readers of stdout
to never receive an error on their reader.

Fx this by tearing down standard streams before we deliver the signal so
that read errors on stdout are propagated properly.

Changelog: fixed