Implement a test helper to simulate context timeout
In some components, handling timeout is essential to their
functionality. Using a context with timeout is the most native and
elegant approach. Unfortunately, context timeout is not friendly to
tests. It's not possible to force a context to expire early and raise
the specific context.DeadlineExceeded error. We can cancel the context,
but the returned error is different and might imply a totally different
meaning.

Thus, in tests, we usually apply a short time.Sleep or use a manual
ticker to simulate a timeout event. Usually, this approach changes the
desired interfaces and makes them more complicated.

This commit implements a test helper to simulate the context timeout. It
creates a wrapper over context.WithCancelCause(). When the timeout event
is triggered, the cancel function dispatches an internal error. The
wrapper captures this error exchange for context.DeadlineExceeded
correspondingly.