limiter: Replace limiter.ConcurrencyLimiter's ticker by context timeout
`limiter.ConcurrencyLimiter` initializer receives a function that
creates a new ticker. This ticker is used to notify the timeout event of
an operation waiting in the queue for too long. This usage pattern is a
bit far-stretched. Using a timeout context is more elegant. The only
reason we used that pattern was for testing purposes, which was
test-friendly.

Thankfully, in a prior commit, a test helper for simulating context
timeout events was introduced. This commit replaces the ticker generator
by context timeout.