limiter: Generalize concurrency limiter's buffers used for semaphore
The concurrency limiter controls the concurrency and queueing mechanism
using two buffered channels. Naturally, we can consider them as two
weighted semaphores. This commit extracts the core operations of those
two semaphores to a staticSemaphore struct. This extraction also
standardizes the interface so that the forward commits can swap in
another semaphore implementation.