Store the repository's relative path in the log entry
The TransactionManager will soon support multiple repositories in
a single partition. Right now there are assumptions in multiple places
that the TransactionManager only ever operates on a single repository.
One of these places is the LogEntry where we don't store the repository
these changes should target. Given the changes may target any of the
repositories in the partition, we need to store the relative path of
the repository to specify the target. This commit does so, and includes
the relative path in the log entry when committing it.