Add a test database hook before reading applied log index
This commit adds a database hook in the tests to invoke code before
the applied log index is read from the database. This occurs when
TransactionManager is initializing. We'll need this to refactor the
TestTransactionManager/failing_initialization_prevents_transaction_beginning
to run the test case without relying on breaking the repository's
state on the disk as we'll soon stop reading it on initialization.