Determine repository existence lazily
TransactionManager currently assumes it operates on a single
repository. This is about to change as partitions need to support
multiple repositories to accommodate for object pools and their
members.

Repository existence checking is one location where this assumption
is made. The repository's existence is determined when initializing
and kept up to date as transactions create and delete the repository.
This becomes more difficult with multiple repositories. We don't want
to check the existence of possibly thousands of repositories in the
partition when initialzing for a read from a single repository. Holding
the state in memory also becomes more complicated.

This commit solves the problem by figuring out whether a repository
exists by checking for it in the filesystem when needed.

This only modifies how this is managed internally in TransactionManager.
External behavior should be unchanged.