Test initialization failures without relying on invalid state
TestTransactionManager/failing_initialization_prevents_transaction_beginning
is testing initialization failures by replacing the repository with a file.
This causes the repository existence check to fail which allows for testing
the initialization failure. This commit updates to trigger the initialization
failure through a panic. This is done as we'll soon support multiple
repositories in a partition as this is necessary for object pool support.
If we have a thousand repositories in a partition, we don't want to check
their existence eagerly when initializing. The code path that is currently
used will thus soon be removed.