Remove repositoryPath field from TransactionManager
The TransactionManager will soon learn to operate on multiple
repositories in a partition. It currently contains a repositoryPath
field that is used in various locations. As a step towards multi-repo
support, remove this field and replace it with getAbsolutePath that
can be used to compute a repository's location based on its relative
path.