Remove repository from TransactionManager
TransactionManager's repository field is a localrepo instance of the
repository the manager is operating on. As we'll have to support
multiple repositories in a partition, we can't rely on such a field
but must build a localrepo instance for the repository we're operating
on. As a step towards that direction, remove the repository field from
TransactionManager and build a localrepo where we need one. Currently
we still use the relativePath stored in the manager but we'll soon pipe
the relative path of the repository with the transaction.