Apply log entry to the correct repository
The TransactionManager is currently applying a log entry always
against the same repository. We're soon about to support multiple
repositories in a partition and we're storing the relative path of
the target repository in the log entry already. This commit changes
ApplyLogEntry to target the application against the relative path
stored in the log entry.