Remove packed-refs locks from the correct repository
Possible stale locks covering 'packed-refs' are currently always cleaned
from a given repository in the TransactionManager. As we need to support
multiple repositories in a partition, this commit changes the lock clean
up to take the relative path from the localrepo instance. That way the
clean up doesn't rely on the relativePath field of the manager as we
move towards removing it.