Inject partition ID into TransactionManager
The TransactionManager is currently using the relative path of the
repository it is operating on to build database keys. This won't
work in the near future as we'll have to support multiple repositories
in a partition. Inject the partition ID into the TransactionManager so
we can use it to later build the database keys in a partition specific
manner without relying on a given repository's relative path.