Use partition ID to construct database keys
TransactionManager is currently expecting to operate on a single
repository and is using its relative path as database key. As the
partitions need to support multiple repositories, this will have
to change. This commit uses the partition's ID to construct its
database keys. The state the TransactionManager is currently storing
in the database is partition specific, not repository specific.