Include the target repository in transaction
The TransactionManager will soon support multiple repositories in a
partition. While we've removed reliance on a given relative path in
most places, there are still few location left. This commit adds a
relative path field in the transaction in order store the target
repository of a transaction. The relativePath field of the manager
is now only used to set the target repository of the transaction.

Internally the manager would now support running transactions
targeting different repositories in the partition. Its Begin()
method however still doesn't allow for specifying the target
repsoitory but we're now set up to make that possible.