Assert partition's state also when repository does not exist
TransactionManager's tests are asserting the content of the partition's
state directory only when the repository exists. This was necessary
before as the TransactionManager stored the WAL state in the repository
itself, and thus it got removed when the repository was deleted. These
days we are storing the state properly in a separate directory independent
of any repository in the partition. We should assert the content of that
directory always regardless of whether any repository exists in the
partition.