Add a helper for asserting state of all repositories in a storage
TransactionManager's tests have a helper for asserting the state of
a single repository. As we are about to support multiple repositories
in a partition, we should assert that the storage contains only expected
repositories and that all of their states match what we expect. This
commit adds a helper for doing such an assertion.